fix(gradle): fix gradle dependencies error
