+++ app/src/fcm/java/org/flyve/mdm/agent/policies/DeployFilePolicy.java
@@ -2,16 +2,14 @@ package org.flyve.mdm.agent.policies;
 
 import android.content.Context;
 
-import org.flyve.mdm.agent.PoliciesController;
 import org.flyve.mdm.agent.utils.FlyveLog;
-import org.flyve.policies.manager.CustomPolicies;
 import org.json.JSONObject;
 
-public class RemoveAppPolicy extends BasePolicies {
+public class DeployFilePolicy extends BasePolicies {
 
-    public static final String POLICY_NAME = "removeApp";
+    public static final String POLICY_NAME = "deployFile";
 
-    public RemoveAppPolicy(Context context) {
+    public DeployFilePolicy(Context context) {
         super(context, POLICY_NAME);
     }
 
@@ -21,13 +19,16 @@ public class RemoveAppPolicy extends BasePolicies {
             JSONObject jsonObj = new JSONObject(message);
 
             if(jsonObj.has(POLICY_NAME)) {
-                String removeApp = jsonObj.getString(POLICY_NAME);
+                String deployFile = jsonObj.getString(POLICY_NAME);
+                String id = jsonObj.getString("id");
+                String versionCode = jsonObj.getString("version");
                 String taskId = jsonObj.getString("taskId");
 
                 // execute the policy
                 PoliciesController policiesController = new PoliciesController(context);
-                policiesController.removePackage(taskId, removeApp);
+                policiesController.downloadFile(deployFile, id, versionCode, taskId);
             }
+
             return true;
         } catch (Exception ex) {
             FlyveLog.e(this.getClass().getName() + ", process", ex.getMessage());
