+++ app/src/fcm/java/org/flyve/mdm/agent/policies/RemoveAppPolicy.java
@@ -2,15 +2,14 @@ package org.flyve.mdm.agent.policies;
 
 import android.content.Context;
 
-import org.flyve.mdm.agent.PoliciesController;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.json.JSONObject;
 
-public class DeployFilePolicy extends BasePolicies {
+public class RemoveAppPolicy extends BasePolicies {
 
-    public static final String POLICY_NAME = "deployFile";
+    public static final String POLICY_NAME = "removeApp";
 
-    public DeployFilePolicy(Context context) {
+    public RemoveAppPolicy(Context context) {
         super(context, POLICY_NAME);
     }
 
@@ -20,16 +19,13 @@ public class DeployFilePolicy extends BasePolicies {
             JSONObject jsonObj = new JSONObject(message);
 
             if(jsonObj.has(POLICY_NAME)) {
-                String deployFile = jsonObj.getString(POLICY_NAME);
-                String id = jsonObj.getString("id");
-                String versionCode = jsonObj.getString("version");
+                String removeApp = jsonObj.getString(POLICY_NAME);
                 String taskId = jsonObj.getString("taskId");
 
                 // execute the policy
                 PoliciesController policiesController = new PoliciesController(context);
-                policiesController.downloadFile(deployFile, id, versionCode, taskId);
+                policiesController.removePackage(taskId, removeApp);
             }
-
             return true;
         } catch (Exception ex) {
             FlyveLog.e(this.getClass().getName() + ", process", ex.getMessage());
